Reestruração das pastas e implemntação de navBar estatica para paginas
