+++ package.json
@@ -17,6 +17,7 @@
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
     "@vitejs/plugin-react": "^3.1.0",
+    "eslint": "^8.37.0",
     "typescript": "^4.9.3",
     "vite": "^4.2.0"
   }
