+++ src/App.tsx
@@ -1,18 +1,30 @@
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import { HomePage } from "./components/HomePage";
-import { UserPage } from "./components/UserPage";
-import { Login } from "./components/Login";
-import { PasswordRecover } from "./components/PasswordRecover";
-import { SingUp } from "./components/SignUp";
+import { HomePage } from "./pages/HomePage/HomePage";
+import { UserPage } from "./pages/UserPage/UserPage";
+import { LoginPage } from "./pages/LoginPage/LoginPage";
+import { PasswordRecoverPage } from "./pages/PasswordRecoveryPage/PasswordRecoverPage";
+import { SignUpPage } from "./pages/SignUpPage/SignUpPage";
+import { RootLayout } from "./pages/Root";
+import { ErrorPage } from "./pages/ErrorPage/ErrorPage";
 
 import "./App.css";
 
 const router = createBrowserRouter([
-  { path: "/", element: <HomePage /> },
-  { path: "/user", element: <UserPage /> },
-  { path: "/login", element: <Login /> },
-  { path: "/password-recover", element: <PasswordRecover /> },
-  { path: "/singup", element: <SingUp /> },
+  {
+    path: "/",
+    element: <RootLayout />,
+    errorElement: <ErrorPage />,
+    children: [
+      { path: "/", element: <HomePage /> },
+      { path: "/user", element: <UserPage /> },
+      { path: "/login", element: <LoginPage /> },
+      { path: "/password-recover", element: <PasswordRecoverPage /> },
+    ],
+  },
+  {
+    path: "/signup",
+    element: <SignUpPage />,
+  },
 ]);
 
 function App() {
