--- src/components/Login.tsx
@@ -1,4 +1,0 @@
-import styles from "./Login.module.css";
-export function Login() {
-  return <div>Login</div>;
-}
