--- src/components/MainNavigation.tsx
@@ -1,5 +1,0 @@
-import styles from "./MainNavifation.module.css";
-
-export function MainNavigation() {
-  return <header></header>;
-}
