+++ src/components/MainNavigation/index.tsx
@@ -1,0 +1,5 @@
+import styles from "./MainNavigation.module.css";
+
+export function MainNavigation() {
+  return <header>Navigation</header>;
+}
