--- src/components/PasswordRecover.module.css
@@ -1,1 +1,0 @@
-import styles from "./PasswordRecover.module.css";
