--- src/components/PasswordRecover.tsx
@@ -1,8 +1,0 @@
-import styles from "./PasswordRecover.module.css";
-export function PasswordRecover() {
-  return (
-    <div>
-      <h1>Recover</h1>
-    </div>
-  );
-}
