--- src/components/SignUp.tsx
@@ -1,7 +1,0 @@
-export function SingUp() {
-  return (
-    <div>
-      <h2>Cadastrar</h2>
-    </div>
-  );
-}
