+++ src/pages/LoginPage/LoginPage.tsx
@@ -1,0 +1,4 @@
+import styles from "./LoginPage.module.css";
+export function LoginPage() {
+  return <div>Login</div>;
+}
