+++ src/pages/PasswordRecoveryPage/PasswordRecoverPage.tsx
@@ -1,0 +1,8 @@
+import styles from "./PasswordRecoverPage.module.css";
+export function PasswordRecoverPage() {
+  return (
+    <div>
+      <h1>Recover</h1>
+    </div>
+  );
+}
