+++ src/pages/Root.tsx
@@ -1,0 +1,11 @@
+import { Outlet } from "react-router-dom";
+import { MainNavigation } from "../components/MainNavigation";
+
+export function RootLayout() {
+  return (
+    <>
+      <MainNavigation />
+      <Outlet />
+    </>
+  );
+}
