+++ src/pages/SignUpPage/SignUpPage.tsx
@@ -1,0 +1,9 @@
+import styles from "./SignUpPage.module.css";
+
+export function SignUpPage() {
+  return (
+    <div>
+      <h2>Cadastrar</h2>
+    </div>
+  );
+}
